refactor: start breaking down larger files
