+++ src/processing/mod.rs
@@ -1,0 +1,3 @@
+//! Functionality for processing
+
+pub mod integrity;
